requirements.txt
# Core
torch>=2.0
transformers>=4.45
datasets>=2.20
peft>=0.12
trl>=0.11

# Unsloth (instalar via pip en Colab — ver scripts/colab_setup.py)
# unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git

# Data
pandas>=2.0
numpy>=1.26
pyyaml>=6.0
Pillow>=10.0
requests>=2.31

# Evaluation
scikit-learn>=1.4

# Dev
pytest>=8.0
jupyter>=1.0
